// src/api/models/MachineDetailDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PK.PHCalc.OutsourcingDummy
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    ProductionPriceDTO,
    ProductionPriceDTOFromJSON,
    ProductionPriceDTOFromJSONTyped,
    ProductionPriceDTOToJSON,
} from './ProductionPriceDTO';

/**
 * 
 * @export
 * @interface MachineDetailDTO
 */
export interface MachineDetailDTO {
    /**
     * 
     * @type {string}
     * @memberof MachineDetailDTO
     */
    name: string;
    /**
     * 
     * @type {Array<ProductionPriceDTO>}
     * @memberof MachineDetailDTO
     */
    productionPrices: Array<ProductionPriceDTO>;
}

export function MachineDetailDTOFromJSON(json: any): MachineDetailDTO {
    return MachineDetailDTOFromJSONTyped(json, false);
}

export function MachineDetailDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): MachineDetailDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['Name'],
        'productionPrices': ((json['ProductionPrices'] as Array<any>).map(ProductionPriceDTOFromJSON)),
    };
}

export function MachineDetailDTOToJSON(value?: MachineDetailDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'Name': value.name,
        'ProductionPrices': ((value.productionPrices as Array<any>).map(ProductionPriceDTOToJSON)),
    };
}