// src/api/models/ProductionPriceDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PK.PHCalc.OutsourcingDummy
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Bounds,
    BoundsFromJSON,
    BoundsFromJSONTyped,
    BoundsToJSON,
} from './Bounds';
import {
    PrintColorCharge,
    PrintColorChargeFromJSON,
    PrintColorChargeFromJSONTyped,
    PrintColorChargeToJSON,
} from './PrintColorCharge';

/**
 * 
 * @export
 * @interface ProductionPriceDTO
 */
export interface ProductionPriceDTO {
    /**
     * 
     * @type {Bounds}
     * @memberof ProductionPriceDTO
     */
    totalAreaSquareMeterBounds: Bounds;
    /**
     * 
     * @type {number}
     * @memberof ProductionPriceDTO
     */
    price: number;
    /**
     * 
     * @type {number}
     * @memberof ProductionPriceDTO
     */
    expressCharge: number;
    /**
     * 
     * @type {Array<PrintColorCharge>}
     * @memberof ProductionPriceDTO
     */
    printColorCharges: Array<PrintColorCharge>;
    /**
     * 
     * @type {boolean}
     * @memberof ProductionPriceDTO
     */
    addPurchasePrice: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ProductionPriceDTO
     */
    addFreightPrice: boolean;
}

export function ProductionPriceDTOFromJSON(json: any): ProductionPriceDTO {
    return ProductionPriceDTOFromJSONTyped(json, false);
}

export function ProductionPriceDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProductionPriceDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'totalAreaSquareMeterBounds': BoundsFromJSON(json['TotalAreaSquareMeterBounds']),
        'price': json['Price'],
        'expressCharge': json['ExpressCharge'],
        'printColorCharges': ((json['PrintColorCharges'] as Array<any>).map(PrintColorChargeFromJSON)),
        'addPurchasePrice': json['AddPurchasePrice'],
        'addFreightPrice': json['AddFreightPrice'],
    };
}

export function ProductionPriceDTOToJSON(value?: ProductionPriceDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'TotalAreaSquareMeterBounds': BoundsToJSON(value.totalAreaSquareMeterBounds),
        'Price': value.price,
        'ExpressCharge': value.expressCharge,
        'PrintColorCharges': ((value.printColorCharges as Array<any>).map(PrintColorChargeToJSON)),
        'AddPurchasePrice': value.addPurchasePrice,
        'AddFreightPrice': value.addFreightPrice,
    };
}